requirements.txt
rdkit-pypi>=2023.9.1
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
openpyxl>=3.1.0
xlrd>=2.0.1
requests>=2.31.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0 